+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/KotlinExtension.kt
@@ -7,6 +7,7 @@ import com.javiersc.hubdle.extensions._internal.PluginIds
 import com.javiersc.hubdle.extensions._internal.state.hubdleState
 import com.javiersc.hubdle.extensions.kotlin.android.KotlinAndroidExtension
 import com.javiersc.hubdle.extensions.kotlin.gradle.KotlinGradleExtension
+import com.javiersc.hubdle.extensions.kotlin.intellij.IntellijExtension
 import com.javiersc.hubdle.extensions.kotlin.jvm.KotlinJvmExtension
 import com.javiersc.hubdle.extensions.kotlin.multiplatform.KotlinMultiplatformExtension
 import javax.inject.Inject
@@ -26,36 +27,31 @@ constructor(
 
     @HubdleDslMarker
     public fun Project.android(action: Action<in KotlinAndroidExtension> = Action {}) {
-        configAndroid(action)
+        action.execute(android)
     }
 
     private val gradle: KotlinGradleExtension = objects.newInstance()
 
     @HubdleDslMarker
     public fun Project.gradle(action: Action<in KotlinGradleExtension> = Action {}) {
-        configGradle(action)
+        action.execute(this@KotlinExtension.gradle)
+    }
+
+    private val intellijPlugin: IntellijExtension = objects.newInstance()
+
+    @HubdleDslMarker
+    public fun Project.intellijPlugin(action: Action<in IntellijExtension> = Action {}) {
+        project.pluginManager.apply(PluginIds.JetBrains.intellij)
+        hubdleState.kotlin.target = intellijPlugin.jvmVersion
+        intellijPlugin.isEnabled = true
+        action.execute(intellijPlugin)
+        hubdleState.kotlin.intellij.isEnabled = intellijPlugin.isEnabled
     }
 
     private val jvm: KotlinJvmExtension = objects.newInstance()
 
     @HubdleDslMarker
     public fun Project.jvm(action: Action<in KotlinJvmExtension> = Action {}) {
-        configJvm(action)
-    }
-
-    private val multiplatform: KotlinMultiplatformExtension = objects.newInstance()
-
-    @HubdleDslMarker
-    public fun Project.multiplatform(action: Action<in KotlinMultiplatformExtension> = Action {}) {
-        configMultiplatform(action)
-    }
-
-    // Configurations
-    private fun Project.configAndroid(action: Action<in KotlinAndroidExtension>) {
-        action.execute(android)
-    }
-
-    private fun Project.configJvm(action: Action<in KotlinJvmExtension>) {
         pluginManager.apply(PluginIds.Kotlin.jvm)
         jvm.isEnabled = true
         action.execute(jvm)
@@ -63,11 +59,10 @@ constructor(
         hubdleState.kotlin.target = jvm.jvmVersion
     }
 
-    private fun Project.configGradle(action: Action<in KotlinGradleExtension>) {
-        action.execute(this@KotlinExtension.gradle)
-    }
+    private val multiplatform: KotlinMultiplatformExtension = objects.newInstance()
 
-    private fun Project.configMultiplatform(action: Action<in KotlinMultiplatformExtension>) {
+    @HubdleDslMarker
+    public fun Project.multiplatform(action: Action<in KotlinMultiplatformExtension> = Action {}) {
         project.pluginManager.apply(PluginIds.Kotlin.multiplatform)
         multiplatform.isEnabled = true
         action.execute(multiplatform)